requirements.txt
flask
google-generativeai
python-dotenv
requests
gspread
google-auth
plotly
